add factory, parser, refactor parser

+++ src/main/java/com/borikov/task1/entity/Point.java
@@ -7,7 +7,7 @@ public class Point {
     private double x;
     private double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         pointId = IdGenerator.generateId();
         this.x = x;
         this.y = y;
